# docs/design/api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "autohunt API",
    "description": "autohunt 系统对外 API 契约（技术设计 v1.3 §3）。\n\n通用约定：JSON；错误统一信封 `{\"error\": {\"code\", \"message\", \"details?\"}}`；时间一律 RFC3339（UTC 存储、前端本地化展示）；列表分页 `?cursor=&limit=`（默认 50）。\n\n鉴权（§3.1）：Web UI 使用 HttpOnly Cookie session（UISession）；外部 Agent 使用 `Authorization: Bearer ah_live_<random>`（AgentBearer）。同一套路由双鉴权，中间件按凭证类型打标 caller ∈ {ui, agent} 供状态机来源裁决（BR-11）。\n\n错误码表：UNAUTHORIZED(401)、FORBIDDEN(403)、NOT_FOUND(404)、PERMIT_REQUIRED(403)/PERMIT_INVALID(403)（submit_token 缺失或无效）、STATE_CONFLICT(409)（状态机裁决拒绝）、VALIDATION_ERROR(422)。\n\n契约 v2 增补（技设 §3.7，M3–M5 写侧，全部仅 UI session，事件详情只读除外）：简历上传/版本管理（/resumes，FR-1/2/3）、档案写（PUT /profile，FR-2）、邮箱账户绑定/解绑/状态/重授权（/email-accounts，FR-40/44）、事件确认/丢弃/修正与原文回溯（/events/{id}/confirm|discard|raw，FR-42/43，BR-2）、通知列表（/notifications，FR-32）、统计与 CSV 导出（/stats/*，FR-50/51/52，口径 §10.4）。\n\n契约 v2 修订（0.2.1）：确认流补 4 项——GET /confirmations 列表（仅 UI）、GET /confirmations/{id} 待确认变体按 caller 区分（UI 返回快照，Agent 仍仅 status）、POST /confirmations/{id}/close 手动关闭（仅 UI）、已确认响应含提交结果回写（submit_result/fail_reason/submitted_at，FR-24）；D-05 读侧三端点（/applications/{id}/history|confirmations|emails，FR-31/24/43）；提醒偏好 GET/PUT /settings/reminders（仅 UI）；GET /applications 增可选 from/to 筛选。既有 v1 冻结 19 端点语义未变动。",
    "version": "0.2.1"
  },
  "paths": {
...
    "/api/v1/stats/overview": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "关键指标卡（FR-52，D-01/D-09 同口径）",
        "description": "筛选参数作用于全部指标（FR-51）。各指标口径见响应字段描述。",
        "operationId": "get_stats_overview_api_v1_stats_overview_get",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按来源渠道筛选（FR-51）",
              "title": "Channel"
            },
            "description": "按来源渠道筛选（FR-51）"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "时间段起（RFC3339，按投递创建时间，FR-51）",
              "title": "From"
            },
            "description": "时间段起（RFC3339，按投递创建时间，FR-51）"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "时间段止（RFC3339）",
              "title": "To"
            },
            "description": "时间段止（RFC3339）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOverview"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/stats/funnel": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "投递漏斗（FR-50，口径 §10.4，AC-7 可核对）",
        "description": "固定四级：已投递 → 笔试 → 面试 → offer；各级 entered_count 与转化率口径见响应字段描述。「待投递」（收藏未投）不计入（§10.4）。筛选参数作用于整页（FR-51）。",
        "operationId": "get_stats_funnel_api_v1_stats_funnel_get",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按来源渠道筛选（FR-51）",
              "title": "Channel"
            },
            "description": "按来源渠道筛选（FR-51）"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "时间段起（RFC3339，按投递创建时间，FR-51）",
              "title": "From"
            },
            "description": "时间段起（RFC3339，按投递创建时间，FR-51）"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "时间段止（RFC3339）",
              "title": "To"
            },
            "description": "时间段止（RFC3339）"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsFunnel"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
    "/api/v1/stats/export": {
      "get": {
        "tags": [
          "stats"
        ],
        "summary": "台账导出 CSV（D-09 明细表导出，本地优先产品的数据可携带性）",
        "description": "导出当前筛选条件下的投递明细；筛选参数与 funnel 一致（FR-51）。",
        "operationId": "export_stats_csv_api_v1_stats_export_get",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按来源渠道筛选（FR-51）",
              "title": "Channel"
            },
            "description": "按来源渠道筛选（FR-51）"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "时间段起（RFC3339，按投递创建时间，FR-51）",
              "title": "From"
            },
            "description": "时间段起（RFC3339，按投递创建时间，FR-51）"
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "时间段止（RFC3339）",
              "title": "To"
            },
            "description": "时间段止（RFC3339）"
          }
        ],
        "responses": {
          "200": {
            "description": "投递台账 CSV（Content-Disposition: attachment; filename=applications-export.csv；UTF-8 带 BOM 便于 Excel 打开）。列：公司/岗位名/渠道/地点/JD 链接/简历版本 ID/投递时间/当前状态/面试轮次/备注（§10.2）。",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED — 未携带有效凭证",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN — 本端点仅接受 UI session 凭证",
            "content": {
              "text/csv": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "UISession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "enum": [
          "待投递",
          "已投递",
          "笔试",
          "面试",
          "offer",
          "已接受",
          "已拒绝",
          "未通过",
          "主动放弃",
          "已过期"
        ],
        "title": "ApplicationStatus",
        "description": "BR-10 状态机状态集。主链 rank 递增；未通过/主动放弃/已过期为旁路终止态。"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHORIZED",
          "FORBIDDEN",
          "NOT_FOUND",
          "PERMIT_REQUIRED",
          "PERMIT_INVALID",
          "STATE_CONFLICT",
          "VALIDATION_ERROR"
        ],
        "title": "ErrorCode"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "统一错误信封（§3 通用约定）。"
      },
...
      "FunnelConversions": {
        "properties": {
          "written_test_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Written Test Rate",
            "description": "笔试转化率 = 进入笔试数 / 已投递及以后数"
          },
          "interview_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interview Rate",
            "description": "面试转化率 = 进入面试数 / 进入笔试数（无笔试环节不剔除）"
          },
          "offer_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offer Rate",
            "description": "offer 转化率 = 进入 offer 数 / 全部已投递数"
          }
        },
        "type": "object",
        "required": [
          "written_test_rate",
          "interview_rate",
          "offer_rate"
        ],
        "title": "FunnelConversions",
        "description": "§10.4 口径；分母为 0 时对应转化率为 null。"
      },
      "FunnelStage": {
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "entered_count": {
            "type": "integer",
            "title": "Entered Count",
            "description": "进入过该状态的投递数：status_history 出现该状态或主链更后状态（去重）；待投递不计入漏斗（§10.4）"
          }
        },
        "type": "object",
        "required": [
          "stage",
          "entered_count"
        ],
        "title": "FunnelStage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatsFunnel": {
        "properties": {
          "stages": {
            "items": {
              "$ref": "#/components/schemas/FunnelStage"
            },
            "type": "array",
            "title": "Stages",
            "description": "固定四级：已投递 → 笔试 → 面试 → offer"
          },
          "conversions": {
            "$ref": "#/components/schemas/FunnelConversions"
          }
        },
        "type": "object",
        "required": [
          "stages",
          "conversions"
        ],
        "title": "StatsFunnel"
      },
      "StatsOverview": {
        "properties": {
          "total_applications": {
            "type": "integer",
            "title": "Total Applications",
            "description": "总投递数：状态≠待投递（与 §10.4 漏斗统计范围一致）"
          },
          "in_progress": {
            "type": "integer",
            "title": "In Progress",
            "description": "进行中：状态 ∈ {已投递, 笔试, 面试, offer}"
          },
          "pending_items": {
            "type": "integer",
            "title": "Pending Items",
            "description": "待确认事项数 = 待确认投递数 + 待确认事件数（D-01 导航红点同口径）"
          },
          "offers": {
            "type": "integer",
            "title": "Offers",
            "description": "offer 数：状态 ∈ {offer, 已接受}（OP-10 仅作数量记录）"
          }
        },
        "type": "object",
        "required": [
          "total_applications",
          "in_progress",
          "pending_items",
          "offers"
        ],
        "title": "StatsOverview",
        "description": "FR-52 指标卡；筛选参数（channel/from/to）作用于全部指标（FR-51）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}